Containerizing QJ Lambda (#8)
* containerizing qj-lambda
